requirements.txt
certifi @ file:///Users/ec2-user/ci_py311/certifi_1678321191472/work/certifi
charset-normalizer==3.1.0
idna==3.4
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15